% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_boxplot.R
\name{generate_boxplot}
\alias{generate_boxplot}
\title{Generate a Boxplot}
\usage{
generate_boxplot(
  data,
  x,
  y,
  color = NULL,
  title = "Boxplot",
  xaxis_title = NULL,
  yaxis_title = NULL
)
}
\arguments{
\item{data}{A data frame to analyze.}

\item{x}{The name of the x variable.}

\item{y}{The name of the y variable.}

\item{color}{The name of the variable for color coding (default: NULL).}

\item{title}{The title of the boxplot (default: 'Boxplot').}

\item{xaxis_title}{The title for the x-axis (default: NULL).}

\item{yaxis_title}{The title for the y-axis (default: NULL).}
}
\value{
Boxplot.
}
\description{
This function generates a boxplot for a given data frame.
}
\examples{
generate_boxplot(mtcars, x = 'gear', y = 'mpg')
}
