% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_confusion_matrix.R
\name{generate_confusion_matrix}
\alias{generate_confusion_matrix}
\title{Generate a Confusion Matrix}
\usage{
generate_confusion_matrix(
  actual,
  predicted,
  labels = NULL,
  title = "Confusion Matrix"
)
}
\arguments{
\item{actual}{A vector of actual values.}

\item{predicted}{A vector of predicted values.}

\item{labels}{A vector of class labels (default: NULL).}

\item{title}{The title of the plot (default: 'Confusion Matrix').}
}
\value{
Confusion matrix plot.
}
\description{
This function generates a confusion matrix for a given set of actual and predicted values.
}
\examples{
actual <- factor(c(1, 0, 1, 1, 0, 1, 0, 0, 1, 0))
predicted <- factor(c(1, 0, 1, 1, 0, 1, 1, 0, 0, 0))
generate_confusion_matrix(actual, predicted)
}
