% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_heatmap.R
\name{generate_heatmap}
\alias{generate_heatmap}
\title{Generate a Heatmap}
\usage{
generate_heatmap(
  data,
  x,
  y,
  fill,
  title = "Heatmap",
  xaxis_title = NULL,
  yaxis_title = NULL
)
}
\arguments{
\item{data}{A data frame to analyze.}

\item{x}{The name of the x variable.}

\item{y}{The name of the y variable.}

\item{fill}{The name of the variable for fill values.}

\item{title}{The title of the heatmap (default: 'Heatmap').}

\item{xaxis_title}{The title for the x-axis (default: NULL).}

\item{yaxis_title}{The title for the y-axis (default: NULL).}
}
\value{
Heatmap plot.
}
\description{
This function generates a heatmap for a given data frame.
}
\examples{
generate_heatmap(mtcars, x = 'gear', y = 'carb', fill = 'mpg')
}
