% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_histogram.R
\name{generate_histogram}
\alias{generate_histogram}
\title{Generate a Histogram}
\usage{
generate_histogram(
  data,
  column,
  bins = 30,
  title = "Histogram",
  xaxis_title = NULL,
  yaxis_title = NULL
)
}
\arguments{
\item{data}{A data frame to analyze.}

\item{column}{The name of the column to plot.}

\item{bins}{Number of bins for the histogram (default: 30).}

\item{title}{The title of the histogram (default: 'Histogram').}

\item{xaxis_title}{The title for the x-axis (default: NULL).}

\item{yaxis_title}{The title for the y-axis (default: NULL).}
}
\value{
Histogram.
}
\description{
This function generates a histogram for a given data frame.
}
\examples{
generate_histogram(mtcars, column = 'mpg')
}
