% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_interactive_bar.R
\name{generate_interactive_bar}
\alias{generate_interactive_bar}
\title{Generate an Interactive Bar Chart}
\usage{
generate_interactive_bar(
  data,
  x,
  y,
  color = NULL,
  title = "Interactive Bar Chart",
  xaxis_title = NULL,
  yaxis_title = NULL
)
}
\arguments{
\item{data}{A data frame to analyze.}

\item{x}{The name of the x variable.}

\item{y}{The name of the y variable.}

\item{color}{The name of the variable for color coding (default: NULL).}

\item{title}{The title of the bar chart (default: 'Interactive Bar Chart').}

\item{xaxis_title}{The title for the x-axis (default: NULL).}

\item{yaxis_title}{The title for the y-axis (default: NULL).}
}
\value{
Interactive bar chart.
}
\description{
This function generates an interactive bar chart using Plotly.
}
\examples{
generate_interactive_bar(mtcars, x = 'gear', y = 'mpg', color = 'cyl')
}
