% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_interactive_line.R
\name{generate_interactive_line}
\alias{generate_interactive_line}
\title{Generate an Interactive Line Chart}
\usage{
generate_interactive_line(
  data,
  x,
  y,
  group = NULL,
  title = "Interactive Line Chart",
  xaxis_title = NULL,
  yaxis_title = NULL
)
}
\arguments{
\item{data}{A data frame to analyze.}

\item{x}{The name of the x variable.}

\item{y}{The name of the y variable.}

\item{group}{The name of the grouping variable (default: NULL).}

\item{title}{The title of the line chart (default: 'Interactive Line Chart').}

\item{xaxis_title}{The title for the x-axis (default: NULL).}

\item{yaxis_title}{The title for the y-axis (default: NULL).}
}
\value{
Interactive line chart.
}
\description{
This function generates an interactive line chart using Plotly.
}
\examples{
generate_interactive_line(mtcars, x = 'hp', y = 'mpg', group = 'cyl')
}
