% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_interactive_pie.R
\name{generate_interactive_pie}
\alias{generate_interactive_pie}
\title{Generate an Interactive Pie Chart}
\usage{
generate_interactive_pie(data, labels, values, title = "Interactive Pie Chart")
}
\arguments{
\item{data}{A data frame to analyze.}

\item{labels}{The name of the variable for pie slice labels.}

\item{values}{The name of the variable for pie slice values.}

\item{title}{The title of the pie chart (default: 'Interactive Pie Chart').}
}
\value{
Interactive pie chart.
}
\description{
This function generates an interactive pie chart using Plotly.
}
\examples{
generate_interactive_pie(mtcars, labels = 'gear', values = 'mpg')
}
