% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_interactive_scatter.R
\name{generate_interactive_scatter}
\alias{generate_interactive_scatter}
\title{Generate an Interactive Scatter Plot}
\usage{
generate_interactive_scatter(
  data,
  x,
  y,
  color = NULL,
  size = NULL,
  title = "Interactive Scatter Plot",
  xaxis_title = NULL,
  yaxis_title = NULL
)
}
\arguments{
\item{data}{A data frame to analyze.}

\item{x}{The name of the x variable.}

\item{y}{The name of the y variable.}

\item{color}{The name of the variable for color coding (default: NULL).}

\item{size}{The name of the variable for size scaling (default: NULL).}

\item{title}{The title of the scatter plot (default: 'Interactive Scatter Plot').}

\item{xaxis_title}{The title for the x-axis (default: NULL).}

\item{yaxis_title}{The title for the y-axis (default: NULL).}
}
\value{
Interactive scatter plot.
}
\description{
This function generates an interactive scatter plot using Plotly.
}
\examples{
generate_interactive_scatter(mtcars, x = 'hp', y = 'mpg', color = 'cyl', size = 'wt')
}
