% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_interactive_table.R
\name{generate_interactive_table}
\alias{generate_interactive_table}
\title{Generate an Interactive Table Using DT}
\usage{
generate_interactive_table(
  data,
  caption = NULL,
  col_names = NULL,
  options = list(pageLength = 5, scrollX = TRUE),
  row_colors = NULL,
  note = NULL,
  footnote = NULL
)
}
\arguments{
\item{data}{A data frame to display.}

\item{caption}{The caption for the table (default: NULL).}

\item{col_names}{Column names for the table (default: NULL).}

\item{options}{List of options for the DT::datatable function (default: list(pageLength = 5, scrollX = TRUE)).}

\item{row_colors}{Row colors for the table (default: NULL).}

\item{note}{Note to add below the table (default: NULL).}

\item{footnote}{Footnote to add below the table (default: NULL).}
}
\value{
Interactive table.
}
\description{
This function generates an interactive table using the DT package.
}
\examples{
generate_interactive_table(mtcars, caption = "Car Dataset")
}
