% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_violin_plot.R
\name{generate_violin_plot}
\alias{generate_violin_plot}
\title{Generate Violin Plot}
\usage{
generate_violin_plot(data, x, y, title = "Violin Plot")
}
\arguments{
\item{data}{A data frame.}

\item{x}{The name of the x variable.}

\item{y}{The name of the y variable.}

\item{title}{Title of the plot.}
}
\value{
A ggplot object of the violin plot.
}
\description{
This function generates a violin plot for a given data frame.
}
\examples{
generate_violin_plot(mtcars, x = "cyl", y = "mpg", title = "Violin Plot of MPG by Cylinder")
}
